Moved Point3D to types file
