+++ industrial_calibration_libs/include/industrial_calibration_libs/targets.h
@@ -1,8 +1,9 @@
 #ifndef TARGETS_H
 #define TARGETS_H
 
+#include <industrial_calibration_libs/types.h>
+
 #include <yaml-cpp/yaml.h>
-#include <cassert>
 #include <iostream>
 #include <memory>
 #include <string>
@@ -16,31 +17,6 @@ enum target_types
   Chessboard = 0,
   CircleGrid = 1,
   ModifiedCircleGrid = 2
-};
-
-struct Point3D
-{
-  Point3D(void);
-
-  Point3D(double x_in, double y_in, double z_in);
-  
-  Point3D(const std::vector<double> &points);
-
-  void setPoints(double x_in, double y_in, double z_in);
-
-  bool setPoints(const std::vector<double> &points);
-
-  std::vector<double> asVector(void);
-
-  bool operator==(const Point3D &p2)
-  {
-    if (this->x == p2.x && this->y == p2.y && this->z == p2.z) {return true;}
-    else {return false;}
-  }
-
-  double x;
-  double y;
-  double z;
 };
 
 struct TargetDefinition
