+++ industrial_calibration_libs/src/types.cpp
@@ -1,0 +1,45 @@
+#include <industrial_calibration_libs/types.h>
+
+namespace industrial_calibration_libs
+{
+Point3D::Point3D(void) { }
+
+Point3D::Point3D(double x_in, double y_in, double z_in) : 
+  x(x_in), y(y_in), z(z_in) { }
+
+Point3D::Point3D(const std::vector<double> &points)
+{
+  assert(points.size() == 3);
+  x = points[0];
+  y = points[1];
+  z = points[2];  
+}
+
+void Point3D::setPoints(double x_in, double y_in, double z_in)
+{
+  x = x_in;
+  y = y_in;
+  z = z_in;
+}
+
+bool Point3D::setPoints(const std::vector<double> &points)
+{
+  if (points.size() != 3) {return false;}
+  else
+  {
+    x = points[0];
+    y = points[1];
+    z = points[2];
+    return true;
+  }
+}
+
+std::vector<double> Point3D::asVector(void)
+{
+  std::vector<double> points;
+  points.resize(3);
+  points[0] = x; points[1] = y; points[2] = z;
+  return points;  
+}
+
+} // namespace industrial_calibration_libs
